+++ api/app_test.go
@@ -2,30 +2,15 @@ package api
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"testing"
-	"time"
 
 	"github.com/spf13/viper"
+	"github.com/stormkit-io/stormkit-cli/model"
 	"github.com/stormkit-io/stormkit-cli/stormkit"
 	"github.com/stormkit-io/stormkit-cli/testutils"
 	"github.com/stretchr/testify/assert"
 )
-
-// expectedApps is mock apps data for tests
-var ExpectedApps = Apps{
-	Apps: []App{
-		{
-			Repo: "repo1",
-			ID:   "1234",
-		},
-		{
-			Repo: "repo2",
-			ID:   "12345",
-		},
-	},
-}
 
 // TestGetApps no app.server parameter
 func TestGetAppsNoServer(t *testing.T) {
@@ -39,7 +24,7 @@ func TestGetAppsNoServer(t *testing.T) {
 // TestGetApps with http code OK
 func TestGetApps(t *testing.T) {
 	// build mock server
-	j, _ := json.Marshal(ExpectedApps)
+	j, _ := json.Marshal(model.MockApps)
 	s := testutils.ServerMock("/apps", j, http.StatusOK)
 	defer s.Close()
 
@@ -50,7 +35,7 @@ func TestGetApps(t *testing.T) {
 
 	// test responses
 	assert.Nil(t, err)
-	assert.Equal(t, &ExpectedApps, apps)
+	assert.Equal(t, &model.MockApps, apps)
 }
 
 func TestGetApps403(t *testing.T) {
@@ -63,13 +48,4 @@ func TestGetApps403(t *testing.T) {
 
 	assert.Nil(t, apps)
 	assert.Contains(t, err.Error(), http.StatusText(http.StatusForbidden))
-}
-
-func TestDumpApp(t *testing.T) {
-	s := DumpApp(ExpectedApps.Apps[0])
-	createdAt := time.Unix(int64(ExpectedApps.Apps[0].CreatedAt), 0)
-	deployedAt := time.Unix(int64(ExpectedApps.Apps[0].DeployedAt), 0)
-	expected := fmt.Sprintf("Repo: repo1\n  ID: 1234\n  Status: false\n  AutoDeploy: \n  DefaultEnv: \n  Endpoint: \n  DisplayName: \n  CreatedAt: %s\n  DeployedAt: %s\n\n", createdAt, deployedAt)
-
-	assert.Equal(t, expected, s)
 }
